A Django-based web application for managing and purchasing watches with features like product catalogs, detailed views, user accounts, and wishlist functionality.
WatchShop is a Django-based web application designed to provide an online platform for browsing and purchasing watches. The application features a product catalog, product details page, user accounts, and a wishlist functionality. It is built using Django to offer a robust and scalable solution for managing an e-commerce store specializing in watches.